// src/generated/models/Cre2b_servicecentersModel.ts
export const Cre2b_servicecenterscre2b_companyfuelavailability = {
  200080000: 'Y',
  200080001: 'N'
} as const;
export type Cre2b_servicecenterscre2b_companyfuelavailability = keyof typeof Cre2b_servicecenterscre2b_companyfuelavailability;
export const Cre2b_servicecenterscre2b_cornetstatus = {
  200080000: 'Y',
  200080001: 'N'
} as const;
export type Cre2b_servicecenterscre2b_cornetstatus = keyof typeof Cre2b_servicecenterscre2b_cornetstatus;
export const Cre2b_servicecenterscre2b_drivercenterstatus = {
  200080000: 'Y',
  200080001: 'N'
} as const;
export type Cre2b_servicecenterscre2b_drivercenterstatus = keyof typeof Cre2b_servicecenterscre2b_drivercenterstatus;
export const Cre2b_servicecenterscre2b_hubpresence = {
  200080000: 'N',
  200080001: 'Y'
} as const;
export type Cre2b_servicecenterscre2b_hubpresence = keyof typeof Cre2b_servicecenterscre2b_hubpresence;
export const Cre2b_servicecenterscre2b_shoppresence = {
  200080000: 'N',
  200080001: 'Y'
} as const;
export type Cre2b_servicecenterscre2b_shoppresence = keyof typeof Cre2b_servicecenterscre2b_shoppresence;
export const Cre2b_servicecentersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_servicecentersstatecode = keyof typeof Cre2b_servicecentersstatecode;
export const Cre2b_servicecentersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_servicecentersstatuscode = keyof typeof Cre2b_servicecentersstatuscode;

export interface Cre2b_servicecentersBase {
  cre2b_centeroperatinghours?: string;
  cre2b_city?: string;
  cre2b_companyfuelavailability?: Cre2b_servicecenterscre2b_companyfuelavailability;
  cre2b_cornetstatus?: Cre2b_servicecenterscre2b_cornetstatus;
  cre2b_district?: string;
  cre2b_drivercenterstatus?: Cre2b_servicecenterscre2b_drivercenterstatus;
  cre2b_facilitymaintenancemanagername?: string;
  cre2b_facilitytype?: string;
  cre2b_hubpresence?: Cre2b_servicecenterscre2b_hubpresence;
  cre2b_localphonenumber?: string;
  cre2b_locationalphacode?: string;
  cre2b_locationname: string;
  cre2b_locationnumber?: number;
  cre2b_operatingcompany?: string;
  cre2b_region?: string;
  cre2b_revenuedoors?: number;
  cre2b_servicecenterid: string;
  cre2b_shoppresence?: Cre2b_servicecenterscre2b_shoppresence;
  cre2b_state?: string;
  cre2b_streetaddress1?: string;
  cre2b_streetaddress2?: string;
  cre2b_supplychainmanagername?: string;
  cre2b_tollfreephonenumber?: string;
  cre2b_totalnumberofdoors?: number;
  cre2b_wocontactname?: string;
  cre2b_wocontactphone?: string;
  cre2b_zipcode?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cre2b_servicecentersstatecode;
  statuscode?: Cre2b_servicecentersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_servicecenters extends Cre2b_servicecentersBase {
  cre2b_companyfuelavailabilityname?: string;
  cre2b_cornetstatusname?: string;
  cre2b_drivercenterstatusname?: string;
  cre2b_hubpresencename?: string;
  cre2b_shoppresencename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}